requirements.txt
kapture==1.1.5
matplotlib==2.1.1
numpy==1.13.3
opencv-python==4.5.4.60
opencv-contrib-python==4.5.4.60
Pillow==9.0.0
PyYAML==6.0
scikit_image==0.17.2
scikit_learn==1.0.2
scipy==0.19.1
skimage==0.0
torch==1.4.0
torchvision==0.11.2
tqdm==4.62.3
